requirements.txt
requests
python-dotenv
pandas
gspread
google-auth
feedparser      # optional (news)
nltk            # optional (VADER sentiment)